pyproject.toml
[project]
name = "notion2md-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "notion2md>=2.9.0",
    "python-dotenv>=1.0.1",
]
